Returns number of intervals per chromosome

Returns number of intervals per chromosome (or chromosome pair).

gintervals.chrom_sizes(intervals = NULL)

Arguments

intervals

intervals set

Value

Data frame representing number of intervals per chromosome (for 1D intervals) or chromosome pair (for 2D intervals).

Details

This function returns number of intervals per chromosome (for 1D intervals) or chromosome pair (for 2D intervals).

See also

Examples

# \dontshow{
options(gmax.processes = 2)
# }

gdb.init_examples()
gintervals.chrom_sizes("annotations")
#>   chrom size
#> 1  chr1    2
#> 2  chr2    6