Returns track attributes values

Returns track attributes values.

gtrack.attr.export(tracks = NULL, attrs = NULL)

Arguments

tracks
a vector of track names or 'NULL'

attrs
a vector of attribute names or 'NULL'

Details

This function returns a data frame that contains track attributes values. Column names of the data frame consist of the attribute names, row names contain the track names.

The list of required tracks is specified by 'tracks' argument. If 'tracks' is 'NULL' the attribute values of all existing tracks are returned.

Likewise the list of required attributes is controled by 'attrs' argument. If 'attrs' is 'NULL' all attribute values of the specified tracks are returned. The columns are also sorted then by "popularity" of an attribute, i.e. the number of tracks containing this attribute. This sorting is not applied if 'attrs' is not 'NULL'.

Empty character string in a table cell marks a non-existing attribute.

Value

A data frame containing track attributes values.

See also

gtrack.attr.import, gtrack.attr.get, gtrack.attr.set

Examples

gdb.init_examples()
gtrack.attr.export()

#>                                                                                                                                                                                                                                                        created.by
#> array_track                                                                                                                                                                                                                                               unknown
#> dense_track         immaculate conception\ngtrack.modify(dense_track, dense_track * 2, intervs)\ngtrack.modify(dense_track, dense_track / 2, intervs)\ngtrack.modify(dense_track, dense_track * 2, intervs)\ngtrack.modify(dense_track, dense_track / 2, intervs)
#> rects_track                                                                                                                                                                                                                                               unknown
#> sparse_track                                                                                                                                                                                    gtrack.create_sparse(sparse_track, description, r, r$test.sparse)
#> subdir.dense_track2                                                                                                                                                               gtrack.create(subdir.dense_track2, description, dense_track * 2, iterator=NULL)
#>                                 created.date
#> array_track         Thu Jun  5 18:09:42 2014
#> dense_track         Tue Dec  9 18:15:21 2014
#> rects_track         Tue Dec  9 18:15:37 2014
#> sparse_track        Tue Jan  7 21:22:31 2014
#> subdir.dense_track2 Wed Jun 11 10:38:22 2014

gtrack.attr.export(tracks = c("sparse_track", "dense_track"))

#>                                                                                                                                                                                                                                                 created.by
#> sparse_track                                                                                                                                                                             gtrack.create_sparse(sparse_track, description, r, r$test.sparse)
#> dense_track  immaculate conception\ngtrack.modify(dense_track, dense_track * 2, intervs)\ngtrack.modify(dense_track, dense_track / 2, intervs)\ngtrack.modify(dense_track, dense_track * 2, intervs)\ngtrack.modify(dense_track, dense_track / 2, intervs)
#>                          created.date
#> sparse_track Tue Jan  7 21:22:31 2014
#> dense_track  Tue Dec  9 18:15:21 2014

gtrack.attr.export(attrs = "created.by")

#>                                                                                                                                                                                                                                                        created.by
#> array_track                                                                                                                                                                                                                                               unknown
#> dense_track         immaculate conception\ngtrack.modify(dense_track, dense_track * 2, intervs)\ngtrack.modify(dense_track, dense_track / 2, intervs)\ngtrack.modify(dense_track, dense_track * 2, intervs)\ngtrack.modify(dense_track, dense_track / 2, intervs)
#> rects_track                                                                                                                                                                                                                                               unknown
#> sparse_track                                                                                                                                                                                    gtrack.create_sparse(sparse_track, description, r, r$test.sparse)
#> subdir.dense_track2                                                                                                                                                               gtrack.create(subdir.dense_track2, description, dense_track * 2, iterator=NULL)