Converts a track to the most current format

Converts a track (if needed) to the most current format.

gtrack.convert(src.track, tgt.track = NULL)

Arguments

src.track
source track name

tgt.track
target track name. If 'NULL' the source track is overwritten.

Details

This function converts a track to the most current format. It should be used if a track created by an old version of the library cannot be read anymore by the newer version. The old track is given by 'src.track'. After conversion a new track 'tgt.track' is created. If 'tgt.track' is 'NULL' the source track is overwritten.

Value

None

See also

gtrack.create, gtrack.2d.create, gtrack.create_sparse