Calculates Wilcoxon test on sliding windows over track expression

Calculates Wilcoxon test on sliding windows over the values of track expression.

gwilcox(expr, winsize1, winsize2, maxpval = 0.05,
        onetailed = TRUE, what2find = 1, intervals = ALLGENOME,
        iterator = NULL, intervals.set.out = NULL)

Arguments

expr
track expression

winsize1
number of values in the first sliding window

winsize2
number of values in the second sliding window

maxpval
maximal P-value

onetailed
if 'TRUE', Wilcoxon test is performed one tailed, otherwise two tailed

what2find
if '-1', lows are searched. If '1', peaks are searched. If '0', both peaks and lows are searched

intervals
genomic scope for which the function is applied

iterator
track expression iterator of "fixed bin" type. If 'NULL' iterator is determined implicitly based on track expression.

intervals.set.out
intervals set name where the function result is optionally outputed

Details

This function runs a Wilcoxon test (also known as a Mann-Whitney test) over the values of track expression in the two sliding windows having an identical center. The sizes of the windows are specified by 'winsize1' and 'winsize2'. 'gwilcox' returns intervals where the smaller window tested against a larger window gives a P-value below 'maxpval'. The test can be one or two tailed.

'what2find' argument controls what should be searched: peaks, lows or both.

If 'intervals.set.out' is not 'NULL' the result is saved as an intervals set. Use this parameter if the result size exceeds the limits of the physical memory.

Value

If 'intervals.set.out' is 'NULL' a data frame representing the intervals with an additional 'pval' column where P-value is below 'maxpval'.

See also

Examples

# NOT RUN {
gdb.init_examples()
gwilcox("dense_track", 100000, 1000, maxpval = 0.01,
        what2find = 1)
# }