Runs R commands on a cluster

Runs R commands on a cluster that supports SGE.

gcluster.run(..., opt.flags = "", max.jobs = 400, debug = FALSE,
             R = "R")

Arguments

...
R commands

opt.flags
optional flags for qsub command

max.jobs
maximal number of simultaneously submitted jobs

debug
if 'TRUE', additional reports are printed

R
command that launches R

Details

This function runs R commands on a cluster by distributing them among cluster nodes. It must run on a machine that supports Sun Grid Engine (SGE). The order in which the commands are executed can not be guaranteed, therefore the commands must be inter-independent.

Optional flags to 'qsub' command can be passed through 'opt.flags' parameter. Users are strongly recommended to use only '-l' flag as other flags might interfere with those that are already used (-terse, -S, -o, -e, -V). For additional information please refer to the manual of 'qsub'.

The maximal number of simultaneously submitted jobs is controlled by 'max.jobs'.

Set 'debug' argument to 'TRUE to allow additional report prints.

'gcluster.run' launches R on the cluster nodes to execute the commands. 'R' argument specifies how R executable should be invoked.

Value

Return value ('retv') is a list, such that 'retv[[i]]' represents the result of the run of command number 'i'. Each result consists of 4 fields that can be accessed by 'retv[[i]]$FIELDNAME':

FIELDNAME 	 DESCRIPTION

exit.status 	 Exit status of the command. Possible values: 'success', 'failure' or 'interrupted'.

retv 	 Return value of the command.

stdout 	 Standard output of the command.

Examples

gdb.init_examples()
v <- 17
gcluster.run(
  gsummary("dense_track + v"),
  {
    intervs <- gscreen("dense_track > 0.1", gintervals(1, 2))
    gsummary("sparse_track", intervs)
  },
  gsummary("rects_track")
)

#> Preparing for distribution...
#> Running the commands...
#> 0%......66%...100%

#> [[1]]
#> [[1]]$exit.status
#> [1] "success"
#> 
#> [[1]]$retv
#> [1] "Error in library(\"misha\", lib.loc = dirname(.GLIBDIR)) : \n  there is no package called ‘misha’\n"
#> attr(,"class")
#> [1] "try-error"
#> attr(,"condition")
#> <simpleError in library("misha", lib.loc = dirname(.GLIBDIR)): there is no package called ‘misha’>
#> 
#> [[1]]$stdout
#> character(0)
#> 
#> [[1]]$stderr
#> [1] "bash: ml: line 1: syntax error: unexpected end of file\nbash: error importing function definition for `BASH_FUNC_ml'\nError in library(\"misha\", lib.loc = dirname(.GLIBDIR)) : \n  there is no package called ‘misha’\n"
#> 
#> 
#> [[2]]
#> [[2]]$exit.status
#> [1] "success"
#> 
#> [[2]]$retv
#> [1] "Error in library(\"misha\", lib.loc = dirname(.GLIBDIR)) : \n  there is no package called ‘misha’\n"
#> attr(,"class")
#> [1] "try-error"
#> attr(,"condition")
#> <simpleError in library("misha", lib.loc = dirname(.GLIBDIR)): there is no package called ‘misha’>
#> 
#> [[2]]$stdout
#> character(0)
#> 
#> [[2]]$stderr
#> [1] "bash: ml: line 1: syntax error: unexpected end of file\nbash: error importing function definition for `BASH_FUNC_ml'\nError in library(\"misha\", lib.loc = dirname(.GLIBDIR)) : \n  there is no package called ‘misha’\n"
#> 
#> 
#> [[3]]
#> [[3]]$exit.status
#> [1] "success"
#> 
#> [[3]]$retv
#> [1] "Error in library(\"misha\", lib.loc = dirname(.GLIBDIR)) : \n  there is no package called ‘misha’\n"
#> attr(,"class")
#> [1] "try-error"
#> attr(,"condition")
#> <simpleError in library("misha", lib.loc = dirname(.GLIBDIR)): there is no package called ‘misha’>
#> 
#> [[3]]$stdout
#> character(0)
#> 
#> [[3]]$stderr
#> [1] "bash: ml: line 1: syntax error: unexpected end of file\nbash: error importing function definition for `BASH_FUNC_ml'\nError in library(\"misha\", lib.loc = dirname(.GLIBDIR)) : \n  there is no package called ‘misha’\n"
#> 
#>