Initializes connection with Genomic Database

Initializes connection with Genomic Database: loads the list of tracks, intervals, etc.

gdb.init(groot, dir = NULL, rescan = FALSE)
gdb.init_examples()

Arguments

groot
the root directory of the Genomic Database

dir
the current working directory inside the Genomic Database

rescan
indicates whether the file structure should be rescanned

Details

'gdb.init' initializes the connection with the Genomic Database. It is typically called first prior to any other function. When the package is attached it internally calls to 'gdb.init.examples' which opens the connection with the database located at 'PKGDIR/trackdb/test' directory, where 'PKGDIR' is the directory where the package is installed.

The current working directory inside the Genomic Database is set to 'dir'. If 'dir' is 'NULL', the current working directory is set to 'GROOT/tracks'.

If 'rescan' is 'TRUE', the list of tracks and intervals is achieved by rescanning directory structure under the current current working directory. Otherwise 'gdb.init' attempts to use the cached list that resides in 'groot/.db.cache' file.

Upon completion the connection is established with the database. If auto-completion mode is switched on (see 'gset_input_method') the list of tracks and intervals sets is loaded and added as variables to the global environment allowing auto-completion of object names with <TAB> key. Also a few global variables are defined. These variables should not be modified by user.

GROOT 	 Root directory of Genomic Database

GWD 	 Current working directory inside Genomic Database

GTRACKS 	 List of all available tracks

GINTERVS 	 List of all available intervals

GVTRACKS 	 List of all available virtual tracks

ALLGENOME 	 List of all chromosomes and their sizes

GITERATOR.INTERVALS 	 A set of iterator intervals for which the track expression is evaluated

Value

None.

See also

gdb.reload, gdb.create, gdir.cd, gtrack.ls, gintervals.ls, gvtrack.ls, gset_input_mode