Creates a track from a track expression

Creates a track from a track expression.

gtrack.create(track, description, expr, iterator = NULL,
              band = NULL)

Arguments

track
track name

description
a character string description

expr
track expression

iterator
track expression iterator. If 'NULL' iterator is determined implicitly based on track expression.

band
track expression band. If 'NULL' no band is used.

Details

This function creates a new track named track. The values of the track are determined by evaluation of 'expr' - a numeric track expression. The type of the new track is determined by the type of the iterator. 'Fixed bin', 'Sparse' or 'Rectangles' track can be created accordingly. 'description' is added as a track attribute.

Value

None.

See also

gtrack.2d.create, gtrack.create_sparse, gtrack.smooth, gtrack.modify, gtrack.rm, gtrack.info, gdir.create

Examples

gdb.init_examples()

## Creates a new track that is a sum of values from 'dense' and
## 2 * non-nan values of 'sparse' track. The new track type is
## Dense with a bin size that equals to '70'.
gtrack.create("mixed_track", "Test track",
              "dense_track +
              replace(sparse_track, is.nan(sparse_track), 0) * 2",
              iterator = 70)
gtrack.info("mixed_track")

#> $type
#> [1] "dense"
#> 
#> $dimensions
#> [1] 1
#> 
#> $size.in.bytes
#> [1] 57160
#> 
#> $bin.size
#> [1] 70
#> 

gtrack.rm("mixed_track", force = T)