Creates a new track from PSSM energy function

Creates a new track from PSSM energy function.

gtrack.create_pwm_energy(track, description, pssmset, pssmid,
                         prior, iterator)

Arguments

track
track name

description
a character string description

pssmset
name of PSSM set: 'pssmset.key' and 'pssmset.data' must be presented in 'GROOT/pssms' directory

pssmid
PSSM id

prior
prior

iterator
track expression iterator for the newly created track

Details

This function creates a new track with values of a PSSM energy function. PSSM parameters (nucleotide probability per position and ploarization) are determined by 'pssmset' key and data files ('pssmset.key' and 'pssmset.data'). These two files must be located in 'GROOT/pssms' directory. The type of the created track is determined by the type of the iterator. 'description' is added as a track attribute.

Value

None.

See also

gtrack.create, gtrack.2d.create, gtrack.create_sparse, gtrack.smooth, gtrack.modify, gtrack.rm, gtrack.info, gdir.create

Examples

gdb.init_examples()
gtrack.create_pwm_energy("pwm_energy_track", "Test track","pssm",
                         3, 0.01, iterator = 100)
gextract("pwm_energy_track", gintervals(1, 0, 1000))

#>    chrom start  end pwm_energy_track intervalID
#> 1   chr1     0  100        -53.42797          1
#> 2   chr1   100  200        -52.64313          1
#> 3   chr1   200  300        -52.11053          1
#> 4   chr1   300  400        -51.71943          1
#> 5   chr1   400  500        -49.75200          1
#> 6   chr1   500  600        -44.75132          1
#> 7   chr1   600  700        -49.12024          1
#> 8   chr1   700  800        -48.81432          1
#> 9   chr1   800  900        -48.37580          1
#> 10  chr1   900 1000        -49.70645          1

gtrack.rm("pwm_energy_track", force = TRUE)