Imports a track from another assembly

Imports a track from another assembly.

gtrack.liftover(track, description, src.track.dir, chain)

Arguments

track
name of a created track

description
a character string description

src.track.dir
path to the directory of the source track

chain
name of chain file or data frame as returned by 'gintervals.load_chain'

Details

This function imports a track located in 'src.track.dir' of another assembly to the current database. Chain file instructs how the conversion of coordinates should be done. It can be either a name of a chain file or a data frame in the same format as returned by 'gintervals.load_chain' function. The name of the newly created track is specified by 'track' argument and 'description' is added as a track attribute.

Value

None.

See also

gintervals.load_chain, gintervals.liftover