Returns a list of track names

Returns a list of track names in Genomic Database.

gtrack.ls(..., ignore.case = FALSE, perl = FALSE, fixed = FALSE,
          useBytes = FALSE)

Arguments

...
these arguments are of either form 'pattern' or 'attribute = pattern'

ignore.case, perl, fixed, useBytes
see 'grep'

Details

This function returns a list of tracks whose name or track attribute value match a pattern (see 'grep'). If called without any arguments all tracks are returned.

If pattern is specified without a track attribute (i.e. in the form of 'pattern') then filtering is applied to the track names. If pattern is supplied with a track attribute (i.e. in the form of 'name = pattern') then track attribute is matched against the pattern.

Multiple patterns are applied one after another. The resulted list of tracks should match all the patterns.

Value

An array that contains the names of tracks that match the supplied patterns.

See also

grep, gtrack.exists, gtrack.create, gtrack.rm

Examples

gdb.init_examples()

# get all track names
gtrack.ls()

#> [1] "array_track"         "dense_track"         "rects_track"        
#> [4] "sparse_track"        "subdir.dense_track2"

# get track names that match the pattern "den*"
gtrack.ls("den*")

#> [1] "dense_track"         "subdir.dense_track2"

# get track names whose "created.by" attribute match the pattern
# "create_sparse"
gtrack.ls(created.by = "create_sparse")

#> [1] "sparse_track"

# get track names whose names match the pattern "den*" and whose
# "created.by" attribute match the pattern "track"
gtrack.ls("den*", created.by = "track")

#> [1] "dense_track"         "subdir.dense_track2"