Returns a list of virtual track names

Returns a list of virtual track names.

gvtrack.ls(pattern = "", ignore.case = FALSE, perl = FALSE,
           fixed = FALSE, useBytes = FALSE)

Arguments

pattern, ignore.case, perl, fixed, useBytes
see 'grep'

Details

This function returns a list of virtual tracks that exist in current R environment that match the pattern (see 'grep'). If called without any arguments all virtual tracks are returned.

Value

An array that contains the names of virtual tracks.

See also

grep, gvtrack.create, gvtrack.rm

Examples

gdb.init_examples()
gvtrack.create("vtrack1", "dense_track", "max")
gvtrack.create("vtrack2", "dense_track", "quantile", 0.5)
gvtrack.ls()

#> [1] "vtrack2" "vtrack3" "vtrack4" "vtrack1"

gvtrack.ls(pattern = "*2")

#> [1] "vtrack2"