Toolkit for analysis of genomic data

'misha' package is intended to help users to efficiently analyze genomic data achieved from various experiments.

For a complete list of help resources, use library(help = "misha").

The following options are available for the package. Use 'options' function to alter the value of the options.

NAME 	 DEFAULT 	 DESCRIPTION

gmax.data.size 	 10000000 	 Maximal number of datums (intervals, ...) in large data sets stored

in memory. Prevents excessive memory usage by various functions

such as 'gextract', 'gscreen', etc.

gbig.intervals.size 	 1000000 	 Minimal number of intervals in a big intervals set format

gmax.mem.usage 	 10000000 	 Maximal memory consumption of all child processes in KB before the limiting algorithm is invoked.

gmax.processes 	 16 	 Maximal number of processes for multitasking

gmax.processes2core 	 2 	 Maximal number of processes per CPU core for multitasking

gmin.scope4process 	 10000 	 Minimal scope range (for 2D: surface) assigned to a

process in multitasking mode.

gbuf.size 	 1000 	 Size of track expression values buffer.

grnd.seed 	 1 	 Seed value for pseudo-random number generator. If '0' seed is set

randomally.

gtrack.chunk.size 	 100000 	 Chunk size in bytes of a 2D track. If '0' chunk size is unlimited.

gtrack.num.chunks 	 0 	 Maximal number of 2D track chunks simultaneously stored in

memory.

More information about the options can be found in 'User manual' of the package.